+++ tensorflow/core/protobuf/autotuning.proto
@@ -1,15 +1,14 @@
-// This file defines protos that store the results of autotuning XLA:GPU
+// This file defines protos that store the results of autotuning various
 // operations.
 //
 // They are in proto format because we want to log them structured. They offer
 // tremendous statistical, testing, and debugging value.
 syntax = "proto3";
 
-package xla.gpu;
+package tensorflow;
 
+import "google/protobuf/any.proto";
 import "google/protobuf/duration.proto";
-import "tensorflow/compiler/xla/xla_data.proto";
-import "tensorflow/compiler/xla/service/hlo.proto";
 
 message CudnnVersion {
   int32 major = 1;
@@ -63,19 +62,12 @@ message AutotuneResult {
   }
 }
 
-message AutotuneLog {
-  message Instruction {
-    xla.HloInstructionProto instruction = 1;
-    repeated xla.ShapeProto operand_shapes = 2;
-  }
-
-  oneof instr_oneof {
-    Instruction instr = 1;
-  }
+message AutotuningLog {
+  google.protobuf.Any instr = 1;
 
   // Records all auto-tuning results per algorithm.
-  repeated AutotuneResult results = 3;
+  repeated AutotuneResult results = 2;
 
-  CudnnVersion cudnn_version = 4;
-  ComputeCapability compute_capability = 5;
+  CudnnVersion cudnn_version = 3;
+  ComputeCapability compute_capability = 4;
 }
